refactor(dt-dashboard): extract subcomponents
