+++ src/pages/DtDashboard.tsx
@@ -13,7 +13,6 @@ import {
   TrendingUp,
   Home,
   Plane,
-  Newspaper,
   Check,
   Trophy,
   Calendar,
@@ -39,8 +38,10 @@ import {
 
 /* ---------- componentes pequeños reutilizados ---------- */
 
-import ProgressBar from '../components/common/ProgressBar';
 import CountdownBar from '../components/common/CountdownBar';
+import SeasonObjectives from './dt-dashboard/SeasonObjectives';
+import LatestNews from './dt-dashboard/LatestNews';
+import QuickActions from './dt-dashboard/QuickActions';
 
 
 /* ---------- componente principal ---------- */
@@ -71,7 +72,6 @@ const DtDashboard = () => {
     possessionDiff(club.id),
     yellowDiff(club.id)
   ];
-  const latestNews = news.slice(0, 3);
 
   return (
     <>
@@ -235,45 +235,10 @@ const DtDashboard = () => {
           </div>
 
           {/* Objetivos de temporada */}
-          <Card className="p-4">
-            <h3 className="mb-4 font-semibold">Objetivos de temporada</h3>
-            <div className="space-y-4">
-              <div>
-                <p className="mb-1 text-xs text-gray-400">Clasificar top 5</p>
-                <ProgressBar value={objectives.position} />
-              </div>
-              <div>
-                <p className="mb-1 text-xs text-gray-400">Juego limpio</p>
-                <ProgressBar value={objectives.fairplay} />
-              </div>
-            </div>
-          </Card>
+          <SeasonObjectives />
 
           {/* Últimas noticias */}
-          <Card className="p-4">
-            <div className="mb-4 flex items-center justify-between">
-              <h3 className="font-semibold">Últimas noticias</h3>
-              <Link
-                to="/liga-master/feed"
-                className="text-accent text-sm hover:underline focus:outline-none focus:ring-2 focus:ring-accent"
-              >
-                Ver todo
-              </Link>
-            </div>
-            <ul className="space-y-3">
-              {latestNews.map(item => (
-                <li key={item.id} className="flex items-center justify-between">
-                  <div className="flex items-center">
-                    <Newspaper size={16} className="mr-2 text-accent" />
-                    <span>{item.title}</span>
-                  </div>
-                  <span className="text-xs text-gray-400">
-                    {formatDate(item.publishDate)}
-                  </span>
-                </li>
-              ))}
-            </ul>
-          </Card>
+          <LatestNews />
         </div>
 
         {/* === COLUMNA DERECHA (1/3) === */}
@@ -336,20 +301,7 @@ const DtDashboard = () => {
           </Card>
 
           {/* Botones de acción rápida */}
-          <div className="grid gap-3 sm:grid-cols-2">
-            <button className="card-hover bg-accent px-4 py-2 font-semibold text-black">
-              Enviar oferta
-            </button>
-            <button className="card-hover bg-accent px-4 py-2 font-semibold text-black">
-              Informe médico
-            </button>
-            <button className="card-hover bg-accent px-4 py-2 font-semibold text-black">
-              Firmar juvenil
-            </button>
-            <button className="card-hover bg-accent px-4 py-2 font-semibold text-black">
-              Publicar declaración
-            </button>
-          </div>
+          <QuickActions />
         </div>
       </main>
