+++ src/pages/dt-dashboard/QuickActions.tsx
@@ -1,0 +1,18 @@
+const QuickActions = () => (
+  <div className="grid gap-3 sm:grid-cols-2">
+    <button className="card-hover bg-accent px-4 py-2 font-semibold text-black">
+      Enviar oferta
+    </button>
+    <button className="card-hover bg-accent px-4 py-2 font-semibold text-black">
+      Informe médico
+    </button>
+    <button className="card-hover bg-accent px-4 py-2 font-semibold text-black">
+      Firmar juvenil
+    </button>
+    <button className="card-hover bg-accent px-4 py-2 font-semibold text-black">
+      Publicar declaración
+    </button>
+  </div>
+);
+
+export default QuickActions;
